refactor(tracimtablelib): migrated everythin tracimTable related to a pseudo lib from frontend_lib
