+++ frontend_lib/src/TracimTableLib/tracimTableLib.js
@@ -1,0 +1,25 @@
+import TracimTable from './TracimTable/TracimTable.jsx'
+import contentFilenameWithBadgesAndBreadcrumbsColumn from './Columns/content/contentFilenameWithBadgesAndBreadcrumbsColumn.jsx'
+import contentInformationColumn from './Columns/content/contentInformationColumn.jsx'
+import contentTypeColumn from './Columns/content/contentTypeColumn.jsx'
+import favoriteButtonColumn from './Columns/favoriteButtonColumn.jsx'
+import timedEventColumn from './Columns/timedEventColumn.jsx'
+import ListItemRowWrapper from './RowWrappers/ListItemRowWrapper.jsx'
+
+const tracimTableLib = {
+  TracimTable,
+  Columns: {
+    content: {
+      contentFilenameWithBadgesAndBreadcrumbsColumn,
+      contentInformationColumn,
+      contentTypeColumn
+    },
+    favoriteButtonColumn,
+    timedEventColumn
+  },
+  RowWrappers: {
+    ListItemRowWrapper
+  }
+}
+
+export default tracimTableLib
