+++ frontend_lib/src/index.js
@@ -304,16 +304,7 @@ export { default as Tag } from './component/Tags/Tag.jsx'
 export { default as Loading } from './component/Loading/Loading.jsx'
 export { default as COLORS } from './colors.js'
 
-export { default as TracimTable } from './component/TracimTable/TracimTable.jsx'
-
-export { default as ListItemRowWrapper } from './container/RowWrappers/ListItemRowWrapper.jsx'
-
-export { default as contentFilenameWithBadgesAndBreadcrumbsColumn } from './columns/content/contentFilenameWithBadgesAndBreadcrumbsColumn.jsx'
-export { default as contentInformationColumn } from './columns/content/contentInformationColumn.jsx'
-export { default as contentTypeColumn } from './columns/content/contentTypeColumn.jsx'
-
-export { default as favoriteButtonColumn } from './columns/favoriteButtonColumn.jsx'
-export { default as timedEventColumn } from './columns/timedEventColumn.jsx'
+export { default as tracimTableLib } from './TracimTableLib/tracimTableLib.js'
 
 export {
   TRANSLATION_STATE,
